deploy.py
import os
import subprocess
from kubernetes import client, config

def setup_kubernetes():
    # Load kubeconfig
    kubeconfig_path = os.getenv("KUBECONFIG_PATH")
    config.load_kube_config(config_file=kubeconfig_path)

    # Create Kubernetes resources
    deployment_yaml = """
    apiVersion: apps/v1
    kind: Deployment
    metadata:
      name: hub
    spec:
      replicas: 1
      selector:
        matchLabels:
          app: hub
      template:
        metadata:
          labels:
            app: hub
        spec:
          containers:
          - name: hub
            image: your-hub-image
            ports:
            - containerPort: 8080
    """

    service_yaml = """
    apiVersion: v1
    kind: Service
    metadata:
      name: hub-service
    spec:
      selector:
        app: hub
      ports:
        - protocol: TCP
          port: 80
          targetPort: 8080
    """

    virtual_service_yaml = """
    apiVersion: networking.istio.io/v1alpha3
    kind: VirtualService
    metadata:
      name: hub-service
    spec:
      hosts:
      - hub-service
      http:
      - route:
        - destination:
            host: hub
            port:
              number: 80
    """

    with open('/tmp/deployment.yaml', 'w') as f:
        f.write(deployment_yaml)
    with open('/tmp/service.yaml', 'w') as f:
        f.write(service_yaml)
    with open('/tmp/virtual_service.yaml', 'w') as f:
        f.write(virtual_service_yaml)

    subprocess.run(["kubectl", "apply", "-f", "/tmp/deployment.yaml"])
    subprocess.run(["kubectl", "apply", "-f", "/tmp/service.yaml"])
    subprocess.run(["kubectl", "apply", "-f", "/tmp/virtual_service.yaml"])

    # Install Istio
    subprocess.run(["istioctl", "install", "--set", "profile=demo", "-y"])

if __name__ == "__main__":
    setup_kubernetes()
