requirements.txt
python-terraform-kubernetes
